Rename integration.rs to guild.rs
